Introduce role: ableton.nix
Like our other roles, this one is designed to install software for
which there is no Debian package nor pre-existing Ansible role. To be
fair, there are actually 2 roles which can install nix, but neither
one has been updated in awhile and I wasn't happy with the
construction of either one.