Rename Mac OS X -> macOS in role metadata
"Mac OS X" was renamed to "macOS" in 2016, which coincides with the
release of 10.13 (Sierra). In some Ansible roles, like
microsoft.azure_modules, they use both platforms in their
meta/main.yml files:

- name: macOS
  versions:
    - Sierra

- name: MacOSX
  versions:
    - 10.11
    - 10.12

However, we don't support macOS <10.13, so for us, it makes more sense
to simply rename this section from "Mac OS X" (which was incorrect for
Ansible Galaxy anyhow, as there shouldn't be spaces in the platform
name) to "macOS", since we can claim to support all of the macOS
versions.