Remove broken `should_install_nix` fact
The `when` clause within the "Set should_install_nix fact" task was
indented, causing it to have no effect (and thus causing
`should_install_nix` to always be true).

Just un-indenting the `when` would then just create an issue later on in
the "Install Nix" task whenever it was undefined.

Instead, remove the extra fact and move its conditional directly to the
(appropriately-indented) `when` clause within "Install Nix".