Use XDG_CONFIG_HOME only if running as nix user
This fixes an issue where Nix packages would fail to be installed if the
playbook was run as a user other than `nix_user`. In such cases, the
xdg_user_config_dir would be propagated with a value for a directory
that the Nix user typically would not have access to.